+++ dart_test.yaml
@@ -1,3 +1,12 @@
 tags:
   analytics:
   documents:
+  dump:
+  exceptions:
+  get_client_stats:
+  get_keys:
+  get_version:
+  health:
+  indexes:
+  search:
+  settings:
